Move CreateComment mutation
