+++ api/queries_comments.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/cli/cli/internal/ghrepo"
 	"github.com/shurcooL/githubv4"
+	"github.com/shurcooL/graphql"
 )
 
 type Comments struct {
@@ -98,4 +99,31 @@ func CommentsForPullRequest(client *Client, repo ghrepo.Interface, pr *PullReque
 	}
 
 	return &Comments{Nodes: comments, TotalCount: len(comments)}, nil
+}
+
+func CommentCreate(client *Client, repoHost string, params map[string]string) (string, error) {
+	var mutation struct {
+		AddComment struct {
+			CommentEdge struct {
+				Node struct {
+					URL string
+				}
+			}
+		} `graphql:"addComment(input: $input)"`
+	}
+
+	variables := map[string]interface{}{
+		"input": githubv4.AddCommentInput{
+			Body:      githubv4.String(params["body"]),
+			SubjectID: graphql.ID(params["subjectId"]),
+		},
+	}
+
+	gql := graphQLClient(client.http, repoHost)
+	err := gql.MutateNamed(context.Background(), "CommentCreate", &mutation, variables)
+	if err != nil {
+		return "", err
+	}
+
+	return mutation.AddComment.CommentEdge.Node.URL, nil
 }
