+++ api/queries_issue.go
@@ -10,7 +10,6 @@ import (
 
 	"github.com/cli/cli/internal/ghrepo"
 	"github.com/shurcooL/githubv4"
-	"github.com/shurcooL/graphql"
 )
 
 type IssuesPayload struct {
@@ -122,33 +121,6 @@ func IssueCreate(client *Client, repo *Repository, params map[string]interface{}
 	}
 
 	return &result.CreateIssue.Issue, nil
-}
-
-func CommentCreate(client *Client, repoHost string, params map[string]string) (string, error) {
-	var mutation struct {
-		AddComment struct {
-			CommentEdge struct {
-				Node struct {
-					URL string
-				}
-			}
-		} `graphql:"addComment(input: $input)"`
-	}
-
-	variables := map[string]interface{}{
-		"input": githubv4.AddCommentInput{
-			Body:      githubv4.String(params["body"]),
-			SubjectID: graphql.ID(params["subjectId"]),
-		},
-	}
-
-	gql := graphQLClient(client.http, repoHost)
-	err := gql.MutateNamed(context.Background(), "CommentCreate", &mutation, variables)
-	if err != nil {
-		return "", err
-	}
-
-	return mutation.AddComment.CommentEdge.Node.URL, nil
 }
 
 func IssueStatus(client *Client, repo ghrepo.Interface, currentUsername string) (*IssuesPayload, error) {
